requirements.txt
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
idna==2.10
iniconfig==1.0.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
